P2PChat.Common

Description

Documentation

data Event

Events used to communicate via Channels

Constructors

CmdInput String
Input from commandline

CmdQuit
Quit from commandline

CmdOutput String
String to print on commandline

SockHostConnect Member [Member]
Client Connect, updated list of members

SockHostDisconnect Member [Member]
Client Disconnect, updates list of member

SockMsgIn String String
User, Msg sock -> main

SockMsgOut String
User, Msg main -> sock

SockClientDisconnect
Client Mode: Socket disconnect

isDisconnect :: Event -> Bool

Checks if the event also forces a disconnect from the server

isQuit :: Event -> Bool

Checks if the event also forces an application quit

data StartMode

Mode to Start in (Host or Client)

Constructors

StartHost

StartClient String Int

data StartConfig

Config to start with

Constructors

StartConfig

Fields

username :: String

hostPort :: Int

mode :: StartMode

data Channels

Communication Channels used for thread communication

Constructors

Channels

Fields

cmain :: Chan Event

cterm :: Chan Event

csock :: Chan Event

data Global

Global Data

Constructors

Global

Fields

myUserName :: String

myUUID :: String

myHostPort :: Int

data Member

Chat Member description

Constructors

Member

Fields

mUsername :: String

mUUID :: String

mHostname :: String

mPort :: Int

data JsonMessage

Generic JsonMessage

Constructors

JsonMessage

Fields

jsType :: String

jsConnect :: Maybe JsonPayloadConnect

jsClientConnected :: Maybe JsonPayloadClientConnected

jsClientDisconnected :: Maybe JsonPayloadClientDisconnected

jsMessage :: Maybe JsonPayloadMessage

data JsonPayloadConnect

JsonPayload when connection

Constructors

JsonPayloadConnect

Fields

jspCname :: String

jspCuuid :: String

jspChostPort :: Int

data JsonPayloadClientConnected

JsonPayload when a client connected

Constructors

JsonPayloadClientConnected

Fields

jspCCmember :: Member

jspCCmembers :: [Member]

data JsonPayloadClientDisconnected

JsonPayload when a client disconnected

Constructors

JsonPayloadClientDisconnected

Fields

jspCDmember :: Member

jspCDmembers :: [Member]

data JsonPayloadMessage

JsonPayload when sending a message

Constructors

JsonPayloadMessage

Fields

jspMname :: String

jspMmsg :: String

jsonEmpty :: String -> JsonMessage

Helper to Construct an empty Message

jsonOK :: JsonMessage

Helper to Construct an Ok message

jsonHeartbeat :: JsonMessage

Helper to Construct a Heartbeat

jsonConnect :: String -> String -> Int -> JsonMessage

Helper to Construct a connect Message

jsonClientConnected :: Member -> [Member] -> JsonMessage

Helper to Construct a ClientConnected Message

jsonClientDisconnected :: Member -> [Member] -> JsonMessage

Helper to Construct a ClientDisconnected Message

jsonMessageSend :: String -> String -> JsonMessage

Helper to Construct a JsonMessage message

jsonParse :: ByteString -> Maybe JsonMessage

Helper to Parse Json from incoming ByteString

isJsonOK :: JsonMessage -> Bool

Helper to to check if its the JsonOK Message